move tmp reader to separate file
